+++ trunk/src/db/db_site.cpp
@@ -461,7 +461,7 @@ Site::findMatches(mstl::string const& name, Matches& result, unsigned maxMatches
 {
 	typedef int (*Compare)(const void *, const void *);
 
-	unsigned n = result.size();
+	size_t n = result.size();
 
 	if (maxMatches <= n)
 		return 0;
@@ -474,7 +474,7 @@ Site::findMatches(mstl::string const& name, Matches& result, unsigned maxMatches
 	::insert(result, i->second);
 	++i;
 
-	unsigned maxSize = maxMatches + n;
+	size_t maxSize = maxMatches + n;
 
 	while (result.size() < maxSize && i != ::siteList.end() && ::isPrefix(name, i->first))
 	{
@@ -482,7 +482,7 @@ Site::findMatches(mstl::string const& name, Matches& result, unsigned maxMatches
 		++i;
 	}
 
-	result.resize(mstl::min(maxMatches + n, result.size()));
+	result.resize(mstl::min(size_t(maxMatches + n), result.size()));
 
 	::qsort(	result.begin(),
 				result.size(),
