+++ trunk/src/mstl/m_byte_buf.cpp
@@ -136,7 +136,7 @@ Codec::compress(unsigned src_bytes, byte const* data)
 			while (p < end && *p == value)
 				++p;
 
-			unsigned rle = mstl::min(16777215, p - src);
+			unsigned rle = mstl::min(size_t(16777215), size_t(p - src));
 
 			if (rle >= Min_RLE_Length)
 			{
@@ -206,7 +206,7 @@ Codec::uncompress(byte_buf const& buf, byte* dst, unsigned dst_bytes)
 	if (*buf.data() == Flag_Copied)
 	{
 		M_ASSERT(dst_bytes + 1 == buf.size());
-		::memcpy(dst, buf.data() + 1, mstl::min(buf.size(), dst_bytes));
+		::memcpy(dst, buf.data() + 1, mstl::min(buf.size(), size_t(dst_bytes)));
 		return;
 	}
